feat: rewrite projects structure & main component
