+++ README.md
@@ -1,70 +1,18 @@
-# Getting Started with Create React App
+# Портал ЭКБ
 
-This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
+Маркетплейс для инженеров.
 
-## Available Scripts
+## Запуск
 
-In the project directory, you can run:
+Проект необходимо скачать. Запуск приложения осуществляется командой: `npm start`.
 
-### `npm start`
+Запускает приложение в режиме разработки. Откройте http://localhost:3000, чтобы просмотреть его в браузере.
 
-Runs the app in the development mode.\
-Open [http://localhost:3000](http://localhost:3000) to view it in your browser.
-
-The page will reload when you make changes.\
-You may also see any lint errors in the console.
-
-### `npm test`
-
-Launches the test runner in the interactive watch mode.\
-See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
-
-### `npm run build`
-
-Builds the app for production to the `build` folder.\
-It correctly bundles React in production mode and optimizes the build for the best performance.
-
-The build is minified and the filenames include the hashes.\
-Your app is ready to be deployed!
-
-See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
-
-### `npm run eject`
-
-**Note: this is a one-way operation. Once you `eject`, you can't go back!**
-
-If you aren't satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
-
-Instead, it will copy all the configuration files and the transitive dependencies (webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you're on your own.
-
-You don't have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn't feel obligated to use this feature. However we understand that this tool wouldn't be useful if you couldn't customize it when you are ready for it.
-
-## Learn More
-
-You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
-
-To learn React, check out the [React documentation](https://reactjs.org/).
-
-### Code Splitting
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/code-splitting](https://facebook.github.io/create-react-app/docs/code-splitting)
-
-### Analyzing the Bundle Size
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size](https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size)
-
-### Making a Progressive Web App
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app](https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app)
-
-### Advanced Configuration
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/advanced-configuration](https://facebook.github.io/create-react-app/docs/advanced-configuration)
-
-### Deployment
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/deployment](https://facebook.github.io/create-react-app/docs/deployment)
-
-### `npm run build` fails to minify
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify](https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify)
+1. к лого добавила курсор и переход на глувную страницу
+2. рефактор структуры проекта, пересоздала папочки, немножечко все передвинула. К каждому компоненту добавила css, не смотря на то, что нет кастомизации отдельной, думаю, стоило бы соблюдать единуб структуру, так легче читать проект
+3. константы теперь в отдельном файле
+4. у меня стоит линтер, поэтому у меня страшные ругательства на то, что в массиве менюхи стоит `target="\_blank"` и якорная заглушка. Это не страшно, но в глазах рябит от ворнингов. Понимаю, что это временное решение, просто хочу пожаловаться
+5. Еще интересный момент, что везде разные кавычки, наверное нужно единообразие
+6. подправила в некоторых местах нейминг и кемел кейс
+7. В мейн в выезжающей менюхе `<a onClick={(e) => e.preventDefault()}>` - ссылка должна обязательно куда-то вести, а тут это просто выезжающая менюха. Я бы тег не `<a></a>` назвала, а `<span></span>`. Зачем там `(e) => e.preventDefault()` тоже не понимаю, мы же не жмем на нее, и это не форма для отправки. Короче я удалила, но может я не поняла задумку, можем вернуть. Даже если мы все же будем кликать на на это, то достаточно он клика, e.preventDefault для форм используется, точнее для отмены стандартного ее поведения
+8. Переименовала ключи в объекте менюхи выезжающей
