--- src/App.css
@@ -1,40 +1,0 @@
-@font-face {
-  font-family: 'InterTightEKB';
-  src: local('InterTightEKB'), url(./fonts/InterTight-Light.ttf) format('truetype');
-  font-weight: 300;
-}
-@font-face {
-  font-family: 'InterTightEKB';
-  src: local('InterTightEKB'), url(./fonts/InterTight-Regular.ttf) format('truetype');
-  font-weight: 400;
-}
-@font-face {
-  font-family: 'InterTightEKB';
-  src: local('InterTightEKB'), url(./fonts/InterTight-Medium.ttf) format('truetype');
-  font-weight: 500;
-}
-@font-face {
-  font-family: 'InterTightEKB';
-  src: local('InterTightEKB'), url(./fonts/InterTight-SemiBold.ttf) format('truetype');
-  font-weight: 600;
-}
-@font-face {
-  font-family: 'InterTightEKB';
-  src: local('InterTightEKB'), url(./fonts/InterTight-Bold.ttf) format('truetype');
-  font-weight: 700;
-}
-@font-face {
-  font-family: 'InterTightEKB';
-  src: local('InterTightEKB'), url(./fonts/InterTight-ExtraBold.ttf) format('truetype');
-  font-weight: 800;
-}
-
-* {
-  margin: 0;
-  padding: 0;
-}
-
-body {
-  font-family: "InterTightEKB", Arial, SansSerif;
-  font-weight: 400;
-}
