--- src/App.js
@@ -1,12 +1,0 @@
-import './App.css';
-import Main from "./main/Main";
-
-function App() {
-  return (
-      <>
-        <Main />
-      </>
-  );
-}
-
-export default App;
