+++ src/components/App/App.css
@@ -1,0 +1,9 @@
+* {
+  margin: 0;
+  padding: 0;
+}
+
+body {
+  font-family: 'InterTightEKB', Arial, SansSerif;
+  font-weight: 400;
+}
