+++ src/components/App/App.js
@@ -1,0 +1,8 @@
+import './App.css';
+import Main from '../Main/Main';
+
+function App() {
+  return <Main />;
+}
+
+export default App;
