+++ src/index.css
@@ -1,0 +1,3 @@
+/* Тут будет подключение normalize и шрифтов */
+@import url('./vendor/normalize.css');
+@import url('./vendor/fonts/fonts.css');
